+++ Core/src/main/java/io/github/bindglam/core/menu/shops/FoodShopMenu.java
@@ -5,14 +5,14 @@ import org.bukkit.Material;
 import java.util.LinkedHashMap;
 
 public class FoodShopMenu extends AbstractShopMenu {
-    public static LinkedHashMap<Material, Integer> items = new LinkedHashMap<>() {{
-        put(Material.COOKED_BEEF, 10);
-        put(Material.COOKED_PORKCHOP, 10);
-        put(Material.COOKED_CHICKEN, 8);
-        put(Material.GOLDEN_CARROT, 100);
-        put(Material.APPLE, 80);
-        put(Material.GOLDEN_APPLE, 800);
-        put(Material.ENCHANTED_GOLDEN_APPLE, 999999);
+    public static LinkedHashMap<Material, Double> items = new LinkedHashMap<>() {{
+        put(Material.COOKED_BEEF, 10.0);
+        put(Material.COOKED_PORKCHOP, 10.0);
+        put(Material.COOKED_CHICKEN, 8.0);
+        put(Material.GOLDEN_CARROT, 100.0);
+        put(Material.APPLE, 80.0);
+        put(Material.GOLDEN_APPLE, 800.0);
+        put(Material.ENCHANTED_GOLDEN_APPLE, 999999.0);
     }};
 
     public FoodShopMenu(){
@@ -20,7 +20,7 @@ public class FoodShopMenu extends AbstractShopMenu {
     }
 
     @Override
-    public LinkedHashMap<Material, Integer> getItems() {
+    public LinkedHashMap<Material, Double> getItems() {
         return items;
     }
 }
