+++ Core/src/main/java/io/github/bindglam/core/menu/shops/FurnitureShopMenu.java
@@ -10,12 +10,12 @@ import org.bukkit.inventory.ItemStack;
 import java.util.LinkedHashMap;
 
 public class FurnitureShopMenu extends AbstractShopMenu {
-    public static LinkedHashMap<Material, Integer> items = new LinkedHashMap<>() {{
+    public static LinkedHashMap<Material, Double> items = new LinkedHashMap<>() {{
     }};
 
-    public static LinkedHashMap<ItemStack, Integer> itemStacks = new LinkedHashMap<>() {{
-        put(CustomFurniture.getInstance("furniture:tv").getItemStack(), 5000000);
-        put(CustomFurniture.getInstance("furniture:chair").getItemStack(), 2000000);
+    public static LinkedHashMap<ItemStack, Double> itemStacks = new LinkedHashMap<>() {{
+        put(CustomFurniture.getInstance("furniture:tv").getItemStack(), 5000000.0);
+        put(CustomFurniture.getInstance("furniture:chair").getItemStack(), 2000000.0);
     }};
 
     public FurnitureShopMenu(){
@@ -23,12 +23,12 @@ public class FurnitureShopMenu extends AbstractShopMenu {
     }
 
     @Override
-    public LinkedHashMap<Material, Integer> getItems() {
+    public LinkedHashMap<Material, Double> getItems() {
         return items;
     }
 
     @Override
-    public LinkedHashMap<ItemStack, Integer> getItemStacks() {
+    public LinkedHashMap<ItemStack, Double> getItemStacks() {
         return itemStacks;
     }
 }
